Updates to demo, remove demove tab-focus code for bare bones implementation
