+++ demo-laerdal/index.html
@@ -46,6 +46,7 @@
       </main>
     </div>
 
-    <script src="dist/demo.js" crossorigin="anonymous"></script>
+    <script src="../dist/plyr.polyfilled.js" crossorigin="anonymous"></script>
+    <script src="./dist/demo.js" crossorigin="anonymous"></script>
   </body>
 </html>
