+++ demo-laerdal/src/sass/settings/colors.scss
@@ -15,7 +15,7 @@ $color-gray-100: lighten($color-gray-200, 9%);
 $color-gray-50: lighten($color-gray-100, 9%);
 
 // Branding
-$color-brand-primary: hsl(198, 100%, 50%);
+$color-brand-primary: #777;
 
 // Text
 $color-text: $color-gray-700;
