+++ demo-laerdal/src/sass/settings/type.scss
@@ -2,8 +2,7 @@
 // Typography
 // ==========================================================================
 
-$font-sans-serif: 'Gordita', 'Avenir', 'Helvetica Neue', sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji',
-  'Segoe UI Symbol';
+$font-sans-serif: 'Lato' sans-serif;
 
 $font-size-base: 15;
 $font-size-small: 13;
